fix(joke): Make the joke perfectly, 100%, safe, always??
As you're building a house, you inevitably have small areas, "voids"
they're called, that are just too challenging to use. You might have
a "void" in the corner of your kitchen counter where a lazy susan
might not quite fit, or you could have one in a corner where some
awkward walls come together.

Despite your best efforts, you'll have them, and when you do, you're
faced with a choice. Do you play is simple: Seal off the void, close
it up, make it dark, and move on? Or do you go to the nearest
Halloween store, find a plastic ghoul or skeleton or similar and
throw it in that void. Maybe you find some fake jewelry, a fake
necklace, perhaps, and drop it in there too. Maybe a real one.

Because someday, somebody, a laborer or the owner — maybe you — is
going to open that void. It might not be this decade or even this
century if you build it right, but someday it will be opened, and that
person deserves a gift. Something to light up their day. And the
choice is yours: Do you put something in there, a joke, or a gem, and
do you make them smile? Or do you proceed on your path, too busy or
too unclever to consider it?

Fixes: #60