Replaced startswith with exact match
The comment above the buggy code indicates that it was trying to account for missing punctuation. I did this by allowing matches where the last character of the canonical form is different/missing.

I also created a test file to make sure Eyecite gets the right court names when fed some of the most common courts.