chore(workflow): Downgrade the python version in the pdoc action
There is a known issue with pdoc3 that prevents it from functioning correctly with Python 3.12. When attempting to execute the command `pdoc module`, the following error occurs: `ModuleNotFoundError: No module named 'distutils'`. As a workaround, temporarily downgrading the python version within the action to `3.11` should allow it to complete successfully.