docs: Simplify hyperscan installation instructions
Back when the hyperscan installation docs were written in
bb741e0daa5d913ae01571f8aecbee188cf087c7 (2021), hyperscan was a bit difficult
to install because it required libhyperscan on the system. But since hyperscan
version 0.3.0 (2022), it has statically linked libhyperscan, so it’s not much
different from other PyPI packages. This commit simplifies the documentation
and drops our CI installation of libhyperscan-dev.